+++ apps/autotest/models/TST/_initialData/010_create_navshortcuts.js
@@ -9,10 +9,9 @@
  * @param {ServerSession} session
  */
 module.exports = function (session) {
-  var
-    desktopID, conn = session.connection
+  let conn = session.connection
 
-  desktopID = conn.lookup('ubm_desktop', 'ID', {expression: 'code', condition: 'equal', values: {code: 'tst_desktop'}})
+  let desktopID = conn.lookup('ubm_desktop', 'ID', {expression: 'code', condition: 'equal', values: {code: 'tst_desktop'}})
   console.info('\tFill `Test` desktop')
   if (!desktopID) {
     console.info('\t\tcreate new `Test` desktop')
@@ -38,7 +37,7 @@ module.exports = function (session) {
       caption: 'Document test',
       iconCls: 'fa fa-folder',
       displayOrder: 10,
-      cmdCode: JSON.stringify({cmdType: 'showList', cmdData: {params: [{ entity: 'tst_document', method: 'select', fieldList: ['favorites.code', 'docDate', 'code', 'description', 'fileStoreSimple']}]}}, null, '\t')
+      cmdCode: JSON.stringify({cmdType: 'showList', cmdData: {params: [{entity: 'tst_document', method: 'select', fieldList: ['favorites.code', 'docDate', 'code', 'description', 'fileStoreSimple']}]}}, null, '\t')
     }
   })
 
@@ -51,7 +50,7 @@ module.exports = function (session) {
       code: 'tst_clob',
       caption: 'CLOB test',
       displayOrder: 20,
-      cmdCode: JSON.stringify({cmdType: 'showList', cmdData: {params: [{ entity: 'tst_clob', method: 'select', fieldList: ['code', 'description', 'text100', 'text2']}]}}, null, '\t')
+      cmdCode: JSON.stringify({cmdType: 'showList', cmdData: {params: [{entity: 'tst_clob', method: 'select', fieldList: ['code', 'description', 'text100', 'text2']}]}}, null, '\t')
     }
   })
 
@@ -69,7 +68,7 @@ module.exports = function (session) {
   })
 
   let displayOrder = 40;
-  ['tst_maindata', 'tst_mainunity', 'tst_dictionary', 'tst_IDMapping', 'tst_histDict'].forEach(function (entityCode) {
+  ['tst_maindata', 'tst_mainunity', 'tst_IDMapping', 'tst_histDict'].forEach(function (entityCode) {
     console.log('\t\t\tcreate `', entityCode, '` shortcut')
     conn.insert({
       fieldList: ['ID'],
@@ -79,12 +78,27 @@ module.exports = function (session) {
         code: entityCode,
         caption: entityCode,
         displayOrder: displayOrder,
-        cmdCode: JSON.stringify({cmdType: 'showList', cmdData: {params: [{ entity: entityCode, method: 'select', fieldList: '*'}]}}, null, '\t')
+        cmdCode: JSON.stringify({cmdType: 'showList', cmdData: {params: [{entity: entityCode, method: 'select', fieldList: '*'}]}}, null, '\t')
       }
     })
     displayOrder = displayOrder + 10
   })
-
+  console.log('\t\t\tcreate `', 'tst_dictionary', '` shortcut')
+  conn.insert({
+    fieldList: ['ID'],
+    entity: 'ubm_navshortcut',
+    execParams: {
+      desktopID: desktopID,
+      code: 'tst_dictionary',
+      caption: 'tst_dictionary',
+      displayOrder: displayOrder,
+      cmdCode: JSON.stringify({cmdType: 'showList',
+        cmdData: {params: [{ entity: 'tst_dictionary',
+          method: 'select',
+          fieldList: ['ID', 'code', 'caption', 'filterValue', 'currencyValue', 'floatValue', 'calculated', 'booleanColumn', 'jsonColumn', 'jsonColumn.propI', 'jsonColumn.propS']}]}}, null, '\t')
+    }
+  })
+  displayOrder = displayOrder + 10
   console.log('\t\t\tcreate `Vue test` shortcut')
   conn.insert({
     fieldList: ['ID'],
@@ -107,7 +121,7 @@ module.exports = function (session) {
       code: 'tst_onlyOffice',
       caption: 'tst_onlyOffice',
       displayOrder: displayOrder,
-      cmdCode: JSON.stringify({cmdType: 'showList', cmdData: {params: [{ entity: 'tst_onlyoffice', method: 'select', fieldList: ['ID', 'caption']}]}}, null, '\t')
+      cmdCode: JSON.stringify({cmdType: 'showList', cmdData: {params: [{entity: 'tst_onlyoffice', method: 'select', fieldList: ['ID', 'caption']}]}}, null, '\t')
     }
   })
 }